Created TwitterScraper with basic functionality
Added code to connect to the Twitter Stream API, filter only allowing tweets pertaining to coca cola, and then process each of those tweets in VaderSentiment, outputting their text and the sentiment analysis of the text.

We need to store this data somehow. I have created a simple little function to base64 the tweets and store them in a simpe text file, however this is not enough for our purposes as we need to timestamp each tweet and be able to grab the aggregate sentiment across a chunk of time. Perhaps SQLite or MySQL?